requirements.txt
faker
flask
flask-caching
flask-compress
flask-cors
flask-limiter
flask-talisman
flask-wtf
gunicorn
httpx
orjson
python-dotenv
validators
werkzeug
pillow